+++ src/app/utils/archive-browser-model.js
@@ -273,7 +273,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
   //#region Action buttons
 
   btnCopyId: computed(function btnCopyId() {
-    return this.createFileAction({
+    return this.createItemBrowserAction({
       id: 'copyArchiveId',
       icon: 'circle-id',
       action: (archives) => {
@@ -295,7 +295,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
     function btnDownloadTar() {
       const disabledTip = this.get('isAnySelectedCreating') ?
         this.t('notAvailableForCreating') : null;
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'downloadTar',
         icon: 'browser-download',
         tip: disabledTip,
@@ -321,7 +321,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
       if (!selectedArchiveHasDip) {
         disabledTip = this.t('selectedArchiveNoDip');
       }
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'browseDip',
         icon: 'browser-directory',
         action: (archives) => {
@@ -339,7 +339,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
 
   btnArchiveProperties: computed(
     function btnArchiveProperties() {
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'archiveProperties',
         icon: 'properties',
         action: (archives) => {
@@ -368,7 +368,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           endingTextInParentheses: this.t('forNonOwnedArchives'),
         });
       }
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'editDescription',
         icon: 'rename',
         tip: disabledTip,
@@ -389,7 +389,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
 
   btnShowAuditLog: computed(
     function btnShowAuditLog() {
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'showAuditLog',
         icon: 'view-list',
         action: (archives) => {
@@ -425,7 +425,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           privilegeFlag: ['space_create_archives'],
         });
       }
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'createArchive',
         icon: 'browser-archive-add',
         tip: disabledTip,
@@ -436,8 +436,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         },
         showIn: [
           actionContext.inDir,
-          actionContext.currentDir,
-          actionContext.spaceRootDir,
+          actionContext.rootDir,
         ],
       });
     }
@@ -481,7 +480,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           });
         }
       }
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'createIncrementalArchive',
         icon: 'browser-archive-add',
         tip: disabledTip,
@@ -530,7 +529,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           privilegeFlag: ['space_recall_archives', 'space_write_data'],
         });
       }
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'recall',
         icon: 'browser-archive-recall',
         tip: disabledTip,
@@ -579,7 +578,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
       } else if (isAnySelectedDeleting) {
         disabledTip = this.t('alreadyDeleting');
       }
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'delete',
         icon: 'browser-delete',
         title: this.t(`fileActions.delete.${areMultipleSelected ? 'multi' : 'single'}`),
@@ -613,7 +612,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           endingTextInParentheses: this.t('forNonOwnedArchives'),
         });
       }
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'cancel',
         icon: 'cancelled',
         title: this.t('fileActions.cancel'),