+++ src/app/utils/filesystem-browser-model.js
@@ -284,7 +284,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           privilegeFlag: 'space_schedule_atm_workflow_executions',
         });
       }
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'bagitUpload',
         class: 'browser-bagit-upload',
         icon: 'browser-archive-upload',
@@ -310,7 +310,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         blockCurrentDirRecalling: true,
       });
       const disabled = Boolean(tip);
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: actionId,
         class: 'browser-upload',
         action: () => uploadManager.triggerUploadDialog(),
@@ -319,7 +319,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         showIn: [
           actionContext.inDir,
           actionContext.currentDir,
-          actionContext.spaceRootDir,
+          actionContext.rootDir,
         ],
       });
     }
@@ -336,7 +336,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         blockCurrentDirRecalling: true,
       });
       const disabled = Boolean(tip);
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: actionId,
         action: () => {
           const {
@@ -350,7 +350,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         showIn: [
           actionContext.inDir,
           actionContext.currentDir,
-          actionContext.spaceRootDir,
+          actionContext.rootDir,
         ],
       });
     }
@@ -391,7 +391,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           privilegeFlag: 'space_manage_shares',
         });
       }
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'share',
         icon: commonActionIcons.shares,
         action: (files) => {
@@ -403,7 +403,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           actionContext.singleFile,
           actionContext.singleDir,
           actionContext.currentDir,
-          actionContext.spaceRootDir,
+          actionContext.rootDir,
         ],
       });
     }
@@ -430,7 +430,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         });
       }
       const disabled = Boolean(disabledTip);
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'datasets',
         icon: 'browser-dataset',
         action: (files) => {
@@ -442,7 +442,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           actionContext.singleDir,
           actionContext.singleFile,
           actionContext.currentDir,
-          actionContext.spaceRootDir,
+          actionContext.rootDir,
         ],
       });
     }
@@ -452,7 +452,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
     const disabledTip = this.generateDisabledTip({
       blockWhenSymlinksOnly: true,
     });
-    return this.createFileAction({
+    return this.createItemBrowserAction({
       id: 'metadata',
       icon: commonActionIcons.metadata,
       disabled: Boolean(disabledTip),
@@ -464,27 +464,27 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         actionContext.singleDir,
         actionContext.singleFile,
         actionContext.currentDir,
-        actionContext.spaceRootDir,
+        actionContext.rootDir,
       ],
     });
   }),
 
   btnInfo: computed(function btnInfo() {
-    return this.createFileAction({
+    return this.createItemBrowserAction({
       id: 'info',
       icon: commonActionIcons.info,
       action: (files, activeTab) => {
         return this.get('openInfo')(files, activeTab);
       },
       showIn: [
-        actionContext.spaceRootDir,
         actionContext.singleDir,
         actionContext.singleFile,
         actionContext.currentDir,
-        actionContext.spaceRootDirPreview,
+        actionContext.rootDir,
         actionContext.singleDirPreview,
         actionContext.singleFilePreview,
         actionContext.currentDirPreview,
+        actionContext.rootDirPreview,
       ],
     });
   }),
@@ -493,7 +493,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
     'showRecallInfoButton',
     function btnRecallInfo() {
       const hidden = !this.get('showRecallInfoButton');
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'recallInfo',
         icon: 'browser-archive-recall',
         action: (files, activeTab) => {
@@ -501,10 +501,10 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         },
         hidden,
         showIn: [
-          actionContext.spaceRootDir,
           actionContext.singleDir,
           actionContext.singleFile,
           actionContext.currentDir,
+          actionContext.rootDir,
         ],
       });
     }
@@ -513,7 +513,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
   btnDownload: computed(
     'selectedItemsContainsOnlyBrokenSymlinks',
     function btnDownload() {
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'download',
         icon: 'browser-download',
         disabled: this.get('selectedItemsContainsOnlyBrokenSymlinks'),
@@ -531,7 +531,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
   btnDownloadTar: computed(
     'selectedItemsContainsOnlyBrokenSymlinks',
     function btnDownloadTar() {
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'downloadTar',
         icon: 'browser-download',
         disabled: this.get('selectedItemsContainsOnlyBrokenSymlinks'),
@@ -539,10 +539,10 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           return this.downloadFiles(files);
         },
         showIn: [
-          actionContext.spaceRootDir,
-          actionContext.spaceRootDirPreview,
           actionContext.currentDir,
           actionContext.currentDirPreview,
+          actionContext.rootDir,
+          actionContext.rootDirPreview,
           actionContext.singleDir,
           actionContext.singleDirPreview,
           actionContext.multiFile,
@@ -557,11 +557,13 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
   ),
 
   btnRename: computed(
+    'isOnlyRootDirSelected',
     'dir.dataIsProtected',
     'selectedItemsContainsRecalling',
     function btnRename() {
       const actionId = 'rename';
       let tip = this.generateDisabledTip({
+        disabledForRootDir: true,
         protectionType: 'data',
         protectionScope: 'final',
         checkProtectionForSelected: false,
@@ -578,7 +580,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         });
       }
       const disabled = Boolean(tip);
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: actionId,
         action: (files) => {
           const {
@@ -593,18 +595,21 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           actionContext.singleDir,
           actionContext.singleFile,
           actionContext.currentDir,
+          actionContext.rootDir,
         ],
       });
     }
   ),
 
   btnPermissions: computed(
+    'isOnlyRootDirSelected',
     'selectedItemsContainsOnlySymlinks',
     function btnPermissions() {
       const disabledTip = this.generateDisabledTip({
+        disabledForRootDir: true,
         blockWhenSymlinksOnly: true,
       });
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'permissions',
         icon: commonActionIcons.permissions,
         disabled: Boolean(disabledTip),
@@ -615,6 +620,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         showIn: [
           ...anySelectedContexts,
           actionContext.currentDir,
+          actionContext.rootDir,
         ],
       });
     }
@@ -622,7 +628,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
 
   btnCreateSymlink: computed('selectedItems.length', function btnCreateSymlink() {
     const areManyFilesSelected = this.get('selectedItems.length') > 1;
-    return this.createFileAction({
+    return this.createItemBrowserAction({
       id: 'createSymlink',
       icon: 'shortcut',
       title: this.t(
@@ -633,7 +639,11 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         browserInstance.setFileClipboardFiles(files.slice());
         browserInstance.setFileClipboardMode('symlink');
       },
-      showIn: anySelectedContexts,
+      showIn: [
+        ...anySelectedContexts,
+        actionContext.currentDir,
+        actionContext.rootDir,
+      ],
     });
   }),
 
@@ -653,7 +663,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           `hardlinkCreatesNewSymlinkTip.${areManyFilesSelected ? 'plural' : 'single'}`
         );
       }
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'createHardlink',
         icon: 'text-link',
         disabled: Boolean(disabledTip),
@@ -666,20 +676,24 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           browserInstance.setFileClipboardFiles(files);
           browserInstance.setFileClipboardMode('hardlink');
         },
-        showIn: anySelectedContexts,
+        showIn: [
+          ...anySelectedContexts,
+          actionContext.currentDir,
+          actionContext.rootDir,
+        ],
       });
     }
   ),
 
   btnPlaceSymlink: computed(function btnPlaceSymlink() {
-    return this.createFileAction({
+    return this.createItemBrowserAction({
       id: 'placeSymlink',
       icon: 'shortcut',
       action: () => this.placeSymlinks(),
       showIn: [
-        actionContext.currentDir,
-        actionContext.spaceRootDir,
         actionContext.inDir,
+        actionContext.currentDir,
+        actionContext.rootDir,
       ],
     });
   }),
@@ -687,14 +701,14 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
   btnPlaceHardlink: computed(
     'fileClipboardFiles.[]',
     function btnPlaceHardlink() {
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'placeHardlink',
         icon: commonActionIcons.hardlinks,
         action: () => this.placeHardlinks(),
         showIn: [
-          actionContext.currentDir,
-          actionContext.spaceRootDir,
           actionContext.inDir,
+          actionContext.currentDir,
+          actionContext.rootDir,
         ],
       });
     }
@@ -704,7 +718,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
     const disabledTip = this.generateDisabledTip({
       blockWhenSymlinksOnly: true,
     });
-    return this.createFileAction({
+    return this.createItemBrowserAction({
       id: 'copy',
       disabled: Boolean(disabledTip),
       tip: disabledTip,
@@ -713,17 +727,23 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         browserInstance.setFileClipboardFiles(files.rejectBy('type', 'symlink'));
         browserInstance.setFileClipboardMode('copy');
       },
-      showIn: anySelectedContexts,
+      showIn: [
+        ...anySelectedContexts,
+        actionContext.currentDir,
+        actionContext.rootDir,
+      ],
     });
   }),
 
   btnCut: computed(
+    'isOnlyRootDirSelected',
     'selectedItemsContainsOnlySymlinks',
     'dir.dataIsProtected',
     'selectedItemsContainsRecalling',
     function btnCut() {
       const actionId = 'cut';
       let tip = this.generateDisabledTip({
+        disabledForRootDir: true,
         protectionType: 'data',
         protectionScope: 'final',
         checkProtectionForSelected: false,
@@ -741,7 +761,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         });
       }
       const disabled = Boolean(tip);
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: actionId,
         action: (files) => {
           const browserInstance = this.get('browserInstance');
@@ -750,7 +770,11 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         },
         disabled,
         tip,
-        showIn: anySelectedContexts,
+        showIn: [
+          ...anySelectedContexts,
+          actionContext.currentDir,
+          actionContext.rootDir,
+        ],
       });
     }
   ),
@@ -767,43 +791,67 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         blockSelectedRecalling: true,
       });
       const disabled = Boolean(tip);
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: actionId,
         action: () => this.pasteFiles(),
         disabled,
         tip,
         showIn: [
-          actionContext.currentDir,
-          actionContext.spaceRootDir,
           actionContext.inDir,
+          actionContext.currentDir,
+          actionContext.rootDir,
         ],
       });
     }
   ),
 
   btnDelete: computed(
+    'isOnlyRootDirSelected',
     'selectedItems.@each.dataIsProtectedByDataset',
     'selectedItemsContainsRecalling',
     function btnDelete() {
       const actionId = 'delete';
       const tip = this.generateDisabledTip({
+        disabledForRootDir: true,
         protectionType: 'data',
         protectionScope: 'dataset',
         blockSelectedRecalling: true,
       });
       const disabled = Boolean(tip);
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: actionId,
-        action: (files) => {
-          const {
-            openRemove,
-            dir,
-          } = this.getProperties('openRemove', 'dir');
-          return openRemove(files, dir);
+        action: async (files) => {
+          const currentDir = this.dir;
+          const isRemovingCurrentDir = files.length === 1 && files[0] === currentDir;
+          let currentDirParent;
+          let onRemoved;
+          if (isRemovingCurrentDir) {
+            currentDirParent = await this.resolveFileParentFun(currentDir);
+            onRemoved = async (removedFiles) => {
+              // check if the dir is still opened as current dir
+              if (
+                removedFiles[0] === currentDir &&
+                currentDir &&
+                this.dir === currentDir &&
+                currentDirParent
+              ) {
+                this.changeDir(currentDirParent);
+              }
+            };
+          }
+          return this.openRemove(
+            files,
+            isRemovingCurrentDir ? currentDirParent : this.dir,
+            onRemoved
+          );
         },
         disabled,
         tip,
-        showIn: anySelectedContexts,
+        showIn: [
+          ...anySelectedContexts,
+          actionContext.currentDir,
+          actionContext.rootDir,
+        ],
       });
     }
   ),
@@ -814,7 +862,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
       const disabledTip = this.generateDisabledTip({
         blockWhenSymlinksOnly: true,
       });
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'distribution',
         icon: commonActionIcons.distribution,
         disabled: Boolean(disabledTip),
@@ -822,7 +870,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
         showIn: [
           ...anySelectedContexts,
           actionContext.currentDir,
-          actionContext.spaceRootDir,
+          actionContext.rootDir,
         ],
         action: (files) => {
           return this.get('openInfo')(files.rejectBy('type', 'symlink'), 'distribution');
@@ -856,13 +904,13 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           privilegeFlag: 'space_view_qos',
         });
       }
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'qos',
         icon: commonActionIcons.qos,
         showIn: [
           ...anySelectedContexts,
           actionContext.currentDir,
-          actionContext.spaceRootDir,
+          actionContext.rootDir,
         ],
         disabled: Boolean(disabledTip),
         tip: disabledTip,
@@ -898,13 +946,13 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
           privilegeFlag: 'space_schedule_atm_workflow_executions',
         });
       }
-      return this.createFileAction({
+      return this.createItemBrowserAction({
         id: 'runWorkflow',
         icon: recordIcon('atmWorkflowSchema'),
         showIn: [
           ...anySelectedContexts,
           actionContext.currentDir,
-          actionContext.spaceRootDir,
+          actionContext.rootDir,
         ],
         disabled: Boolean(disabledTip),
         tip: disabledTip,
@@ -1212,6 +1260,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
   /**
    *
    * @param {Object} options
+   * @param {boolean} [disabledForRootDir]
    * @param {'data'|'metadata'} [protectionType]
    * @param {'final'|'dataset'} [protectionScope] `final` - check final effective
    *   protection of file (concerns datasets and hardlinks inherited protection);
@@ -1225,6 +1274,7 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
    * @returns
    */
   generateDisabledTip({
+    disabledForRootDir,
     protectionType,
     protectionScope = 'final',
     checkProtectionForCurrentDir = false,
@@ -1249,6 +1299,9 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
       return;
     }
     let tip;
+    if (disabledForRootDir && this.isOnlyRootDirSelected) {
+      return this.t('disabledActionReason.notAvailableForRootDir');
+    }
     if (!tip && protectionType) {
       let protectionProperty = `${protectionType}IsProtected`;
       if (protectionScope === 'dataset') {